Create Nimc soln
The National Identification Management Commission (NIMC) is empowered by the constitution to provide unique identification number (now NIN) to citizens of the Federal Republic of Nigeria. Their efforts thus far has help a lot, such that even if you live in Lagos, your information is still the same when you live in Kano, Edo, Anambra, Kogi, or wherever you found yourself. However, there are some lapses in NIMC System. As of the time of posting this task NIMC system is down, under maintenance for days. Accessing Corporate Affair Commission or other commissions to validate your NIN is currently on hold. Imagine if Nigeria was fully relying on this NIMC system for validation, protection and other national security matters! This is one of the lapses I believe such an organization of such magnitude shouldn’t think of having.